requirements.txt
cryptography
python-dotenv
flask
flasgger
flask_staticdirs
pyjwt
jsonpatch
